move copert helper functions to strategy_helpers/helpers.py
